requirements.txt
Django==1.9.4
argparse==1.2.1
djangorestframework==3.3.3
rest-condition==1.0.2
wsgiref==0.1.2
